descriptor tables above 4 GiB, and results that zero extend
A 64-bit kernel builds its gdt and idt in the high half of the address space,
and v86 held both bases in an i32 - the 64-bit lgdt read eight bytes and then
truncated them. That is the instruction a windows boot stops on once everything
before it works. The bases widen to i64 and move to 360 and 368 where an i64 is
aligned; the limits beside them are 16-bit quantities and stay. With them widen
the descriptor addresses the two lookups hand back, and the system reads and
writes that reach the tables.

Writing the fixture for it turned up something worse. In 64-bit mode every
32-bit write to a register clears the upper half, and write_reg32 does not - it
leaves it alone, which is right in 32-bit mode, where there is no upper half to
speak of. cpuid, rdtsc and rdmsr are all delegated to their 32-bit
implementations, so in 64-bit mode each was answering with its result merged
into whatever was in the upper half: not a corrupted corner of the answer but a
completely different, much larger number. Windows reads cpuid before it will
boot.

Fixing write_reg32 itself would be tidier, but the jit writes 32-bit registers
with its own generated stores and would not follow, and the differential
fixtures compare the two - so the three delegating sites zero extend instead.

+++ src/rust/cpu/instructions_0f.rs
@@ -200,7 +200,7 @@ pub unsafe fn instr32_0F01_0_reg(_r: i32) { trigger_ud(); }
 unsafe fn sgdt(addr: i32, mask: i32) {
     return_on_pagefault!(writable_or_pagefault(addr, 6));
     safe_write16(addr, *gdtr_size).unwrap();
-    safe_write32(addr + 2, *gdtr_offset & mask).unwrap();
+    safe_write32(addr + 2, *gdtr_offset as i32 & mask).unwrap();
 }
 #[no_mangle]
 pub unsafe fn instr16_0F01_0_mem(addr: i32) { sgdt(addr, 0xFFFFFF) }
@@ -215,7 +215,7 @@ pub unsafe fn instr32_0F01_1_reg(_r: i32) { trigger_ud(); }
 unsafe fn sidt(addr: i32, mask: i32) {
     return_on_pagefault!(writable_or_pagefault(addr, 6));
     safe_write16(addr, *idtr_size).unwrap();
-    safe_write32(addr + 2, *idtr_offset & mask).unwrap();
+    safe_write32(addr + 2, *idtr_offset as i32 & mask).unwrap();
 }
 #[no_mangle]
 pub unsafe fn instr16_0F01_1_mem(addr: i32) { sidt(addr, 0xFFFFFF) }
@@ -235,7 +235,7 @@ unsafe fn lgdt(addr: i32, mask: i32) {
     let size = return_on_pagefault!(safe_read16(addr));
     let offset = return_on_pagefault!(safe_read32s(addr + 2));
     *gdtr_size = size;
-    *gdtr_offset = offset & mask;
+    *gdtr_offset = (offset & mask) as i64;
 }
 #[no_mangle]
 pub unsafe fn instr16_0F01_2_mem(addr: i32) { lgdt(addr, 0xFFFFFF); }
@@ -255,7 +255,7 @@ unsafe fn lidt(addr: i32, mask: i32) {
     let size = return_on_pagefault!(safe_read16(addr));
     let offset = return_on_pagefault!(safe_read32s(addr + 2));
     *idtr_size = size;
-    *idtr_offset = offset & mask;
+    *idtr_offset = (offset & mask) as i64;
 }
 #[no_mangle]
 pub unsafe fn instr16_0F01_3_mem(addr: i32) { lidt(addr, 0xFFFFFF); }
